# refactor(rome_parser): Move lexer inside parser