--- a/.gitattributes
+++ b/.gitattributes
@@ -1,3 +1,3 @@
 * text=auto eol=lf
-crates/rome_js_lexer/src/tables.rs linguist-generated=true text=auto eol=lf
+crates/rome_js_parser/src/lexer/tables.rs linguist-generated=true text=auto eol=lf
 crates/rome_js_syntax/src/generated/* linguist-generated=true text=auto eol=lf