… Cargo.lock
… crates/rome_js_lexer/CHANGELOG.md
… crates/rome_js_lexer/Cargo.toml
… crates/rome_js_lexer/LICENSE
… crates/rome_js_parser/Cargo.toml
… crates/rome_js_parser/src/lexer/bytes.rs
… crates/rome_js_parser/src/lexer/errors.rs
… crates/rome_js_parser/src/lexer/highlight.rs
… crates/rome_js_parser/src/parser/single_token_parse_recovery.rs
… crates/rome_js_parser/src/syntax/expr.rs
… crates/rome_js_parser/src/syntax/module.rs
… crates/rome_js_parser/src/syntax/program.rs
… crates/rome_js_parser/src/syntax/stmt.rs
… crates/rome_js_parser/src/syntax/typescript/types.rs
… crates/rome_js_parser/src/token_source.rs
… website/src/credits.md
… xtask/bench/src/features/parser.rs
… xtask/codegen/src/unicode.rs